Revert "Feature/auto svr"
