+++ .gitignore
@@ -1,5 +1,4 @@
 venv/
-pyenv3-8/
 
 tests.ipynb
 test_datasets/
@@ -9,5 +8,3 @@ __pycache__/
 
 dist/
 auto_modeling.egg-info/
-
-.pytest_cache
